Revert "[WebConsole] Refactor Log stream for better selection UX, better stick-to-bottom behavior"
This reverts commit 129cbcda69082665c25607a9e054243d6cab9050.
