+++ web-console/package.json
@@ -30,7 +30,6 @@
     "@fortawesome/fontawesome-free": "^6.7.2",
     "@hey-api/client-fetch": "0.1.10",
     "@hey-api/openapi-ts": "^0.47.2",
-    "@humanspeak/svelte-virtual-list": "^0.2.4",
     "@playwright/experimental-ct-svelte": "^1.50.0",
     "@playwright/test": "^1.50.0",
     "@poppanator/sveltekit-svg": "^5.0.0",
@@ -61,7 +60,6 @@
     "eslint": "^9.19.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.46.1",
-    "fancy-ansi": "^0.1.3",
     "fast-deep-equal": "^3.1.3",
     "flowbite-svelte": "^0.46.23",
     "formsnap": "^2.0.0",
@@ -76,11 +74,11 @@
     "prettier": "^3.4.2",
     "prettier-plugin-svelte": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
+    "pretty-ansi": "^3.0.0",
     "runed": "^0.15.4",
     "sass-embedded": "^1.83.4",
     "sk-oidc-oauth": "^1.2.0",
     "sort-on": "^6.1.0",
-    "strip-ansi": "^7.1.0",
     "style-to-object": "^1.0.8",
     "svelte": "^5.19.5",
     "svelte-attr": "^0.0.2",
@@ -101,7 +99,7 @@
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.22.0",
     "valibot": "^1.0.0-beta.14",
-    "virtua": "^0.40.0",
+    "virtua": "^0.36.3",
     "vite": "^5.4.14",
     "vite-svg-2-webfont": "^3.4.0",
     "worker-timers": "^8.0.13"
