+++ web-console/src/lib/functions/pipelines/changeStream.ts
@@ -277,40 +277,3 @@ export const pushAsCircularBuffer =
     arr().push(...vs.slice(0, numNewItems))
     return numItemsToDrop
   }
-
-/**
- *
- * @returns Index offset of items in the original list after push (positive number)
- */
-export const pushAsCircularBuffer2 =
-  <T>(
-    getStr: () => string,
-    appendStr: (...str: string[]) => void,
-    rowBoundaries: () => number[],
-    bufferSize: number,
-    mapValue: (v: T) => string
-  ) =>
-  (values: T[]) => {
-    const vs = values.map(mapValue)
-    const lengths = vs.map((v) => v.length)
-    const deltaLength = lengths.reduce((acc, cur) => acc + cur, 0)
-    const offset = getStr().length + deltaLength - bufferSize
-    let baseBoundaryIndex = rowBoundaries().findIndex((b) => b >= offset) ?? offset // Find a boundary to slice on
-    let baseBoundary = rowBoundaries()[baseBoundaryIndex]
-    rowBoundaries().splice(0, baseBoundaryIndex + 1)
-    rowBoundaries().forEach((_, i) => (rowBoundaries()[i] -= baseBoundary))
-    rowBoundaries().push(...lengths)
-    getStr().slice(0, offset)
-    let deltaBoundary = 0
-    let deltaSliceIndex = lengths.findIndex((b) => {
-      deltaBoundary += b
-      return deltaBoundary > bufferSize
-    })
-    if (deltaSliceIndex === -1) {
-      deltaSliceIndex = lengths.length
-    }
-    deltaSliceIndex = lengths.length - deltaSliceIndex
-    const newRows = vs.slice(deltaSliceIndex)
-    appendStr(...newRows)
-    return Math.max(offset, 0)
-  }
