+++ web-console/src/lib/services/pipelineManager.ts
@@ -340,7 +340,7 @@ export const postPipelineAction = async (
         continue
       }
       throw new Error(
-        `Unexpected status ${JSON.stringify(status)} while waiting for pipeline ${pipeline_name} to complete action ${action}`
+        `Unexpected status ${status} while waiting for pipeline ${pipeline_name} to complete action ${action}`
       )
     }
     return
